% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities.}
\format{A data frame with 123,203 rows and 3 variables:
\describe{
  \item{country}{Country ISO code}
  \item{city}{City name}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
}}
\source{
\url{https://github.com/echarts-maps/echarts-cities-js}
}
\description{
This is a dataset with the latitude and longitude of all cities on planet earth with greater than 1000 people.
}
\keyword{data}
