% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_maps.R
\name{em_maps}
\alias{em_maps}
\title{Maps}
\usage{
em_maps(e, maps, name)
}
\arguments{
\item{e}{An object of class \code{echarts4r}.}

\item{maps}{Map names, a vector of countries.}

\item{name}{the regist map name which can be used in e_map}
}
\description{
Add map dependency to \link[echarts4r]{e_charts} chart. It can combine two or more countries into one plot.
}
\details{
essentially a wrapper to \link[echarts4r]{e_map_register}.
}
\examples{

\dontrun{
maps <- c("China","Japan") 
library(echarts4r)  

e_charts() \%>\% 
   em_maps(maps,"user") \%>\% 
   e_map(map = "user") 
}

}
