% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{maps}
\alias{maps}
\alias{em_map}
\alias{em_bank}
\title{Maps}
\usage{
em_map(e, map)

em_bank()
}
\arguments{
\item{e}{An object of class \code{echarts4r}.}

\item{map}{Map name, a country.}
}
\description{
Add map dependency to \link[echarts4r]{e_charts} chart.
}
\examples{
maps <- em_bank() # list of all maps

library(echarts4r)

df <- data.frame(x = "", y = "")

df \%>\% 
  e_charts(x) \%>\%
  em_map("France") \%>\% 
  e_map(y, mapType = "法国", map = NULL) 

}
