% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{maps}
\alias{maps}
\alias{em_map}
\alias{em_bank}
\alias{get_json}
\title{Maps}
\usage{
em_map(e, map)

em_bank()

get_json(map)
}
\arguments{
\item{e}{An object of class \code{echarts4r}.}

\item{map}{Map name, a country.}
}
\description{
Add map dependency to \link[echarts4r]{e_charts} chart.
}
\details{
essentially a wrapper to \link[echarts4r]{e_map_register}.
}
\examples{
maps <- em_bank() # list of all maps

\dontrun{
library(echarts4r)

df <- data.frame(x = "", y = "")

df \%>\% 
  e_charts(x) \%>\%
  em_map("France") \%>\% 
  e_map(y, map = "france") 
}

}
